remove leftover code
