--- packages/arb-token-bridge-ui/src/components/TransferPanel/TokenInfoTooltip.tsx
@@ -1,75 +1,0 @@
-import { Tooltip } from '@/app/components/common/Tooltip';
-
-import { ERC20BridgeToken } from '../../hooks/arbTokenBridge.types';
-import { useNativeCurrency } from '../../hooks/useNativeCurrency';
-import { useNetworks } from '../../hooks/useNetworks';
-import { useNetworksRelationship } from '../../hooks/useNetworksRelationship';
-import { ChainId } from '../../types/ChainId';
-import { shortenAddress } from '../../util/CommonUtils';
-import { isTokenNativeUSDC } from '../../util/TokenUtils';
-import { getExplorerUrl } from '../../util/networks';
-import { ExternalLink } from '../common/ExternalLink';
-import { TokenLogo } from './TokenLogo';
-
-export function BlockExplorerTokenLink({
-  chainId,
-  address,
-}: {
-  chainId: ChainId;
-  address: string | undefined;
-}) {
-  if (typeof address === 'undefined') {
-    return null;
-  }
-
-  return (
-    <ExternalLink
-      href={`${getExplorerUrl(chainId)}/token/${address}`}
-      className="arb-hover text-xs underline"
-      onClick={(e) => e.stopPropagation()}
-    >
-      {shortenAddress(address).toLowerCase()}
-    </ExternalLink>
-  );
-}
-
-export const TokenInfoTooltip = ({ token }: { token: ERC20BridgeToken | null }) => {
-  const [networks] = useNetworks();
-  const { isDepositMode, childChainProvider } = useNetworksRelationship(networks);
-  const childChainNativeCurrency = useNativeCurrency({
-    provider: childChainProvider,
-  });
-
-  if (!token || isTokenNativeUSDC(token.address)) {
-    return <span>{token?.symbol ?? childChainNativeCurrency.symbol}</span>;
-  }
-
-  const tokenAddress = isDepositMode ? token.l2Address : token.address;
-
-  return (
-    <Tooltip
-      content={
-        <div className="flex items-center space-x-2">
-          <TokenLogo srcOverride={token.logoURI} className="h-7 w-7" />
-          <div className="flex flex-col">
-            <div className="flex space-x-1">
-              <span className="text-lg font-normal">{token.symbol}</span>
-              <span className="pt-[4px] text-xs font-normal text-gray-400">{token.name}</span>
-            </div>
-            <BlockExplorerTokenLink chainId={networks.destinationChain.id} address={tokenAddress} />
-          </div>
-        </div>
-      }
-      tooltipProps={{
-        disableHoverableContent: false,
-      }}
-    >
-      <ExternalLink
-        href={`${getExplorerUrl(networks.destinationChain.id)}/token/${tokenAddress}`}
-        className="cursor-pointer underline"
-      >
-        {token.symbol}
-      </ExternalLink>
-    </Tooltip>
-  );
-};
